[feature] navigation 구조 변경
