+++ presentation/src/main/java/com/daejol/presentation/main/MainRemember.kt
@@ -1,0 +1,16 @@
+package com.daejol.presentation.main
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import com.daejol.presentation.worldcup.WorldCupViewModel
+
+@Composable
+fun worldCupViewModel(): WorldCupViewModel = hiltViewModel()
+
+@Composable
+fun rememberAppState(
+    navController: NavHostController = rememberNavController()
+) = remember(navController) { AppState(navController) }
