+++ presentation/src/main/java/com/daejol/presentation/main/MainScreen.kt
@@ -1,0 +1,68 @@
+package com.daejol.presentation.main
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.daejol.presentation.Screen
+
+@Composable
+fun MainScreen() {
+    val items = listOf(
+        Screen.Home,
+        Screen.Category,
+        Screen.Bookmark,
+        Screen.MyPage,
+    )
+    val appState = rememberAppState()
+    val navController = appState.navController
+
+    Scaffold(
+        bottomBar = {
+            if (appState.shouldShowBottomBar) {
+                NavigationBar {
+                    val navBackStackEntry by navController.currentBackStackEntryAsState()
+                    val currentDestination = navBackStackEntry?.destination
+                    items.forEach { screen ->
+                        NavigationBarItem(
+                            icon = {
+                                screen.icon?.let {
+                                    Icon(
+                                        screen.icon,
+                                        contentDescription = null
+                                    )
+                                }
+                            },
+                            label = { Text(stringResource(screen.resourceId)) },
+                            selected = currentDestination?.hierarchy?.any { it.route == screen.route } == true,
+                            onClick = {
+                                navController.navigate(screen.route) {
+                                    popUpTo(navController.graph.findStartDestination().id) {
+                                        saveState = true
+                                    }
+                                    launchSingleTop = true
+                                    restoreState = true
+                                }
+                            }
+                        )
+                    }
+                }
+            }
+        }
+    ) { innerPadding ->
+        MainNavHost(navController, Modifier.padding(innerPadding))
+    }
+}
+
